rootfs/app/requirements.txt
PyYAML==6.0.3
requests==2.32.5
paho-mqtt==2.1.0
pymodbus==2.5.3
SungrowModbusTcpClient==0.1.6
SungrowModbusWebClient==0.3.2